+++ client/cmdlfawid.c
@@ -333,7 +333,7 @@ static int CmdAWIDDemod(const char *Cmd) {
 
 // this read is the "normal" read,  which download lf signal and tries to demod here.
 static int CmdAWIDRead(const char *Cmd) {
-    lf_read(true, 12000);
+    lf_read(false, 12000);
     return CmdAWIDDemod(Cmd);
 }