+++ client/cmdlfhid.c
@@ -255,7 +255,7 @@ static int CmdHIDDemod(const char *Cmd) {
 
 // this read is the "normal" read,  which download lf signal and tries to demod here.
 static int CmdHIDRead(const char *Cmd) {
-    lf_read(true, 12000);
+    lf_read(false, 12000);
     return CmdHIDDemod(Cmd);
 }